perldelta for most of 5.13.10
Add a perldelta for the commits since v5.13.9 that I thought worthy of
mention.

What should be included was discussed on perl5-porters in the "The
status of the perldelta for 5.13.10" and "[PATCH/RFC]
pod/perldelta.pod: WIP perldelta from v5.13.9..356123f144" threads.

There might still be things that need a delta that I either don't
fully understand or don't have the competency to write a perldelta
for. These were when I started writing this commit message at least:

    $ git-deltatool --mode summary --type blocking
    Scanning for blocking commits since v5.13.9...
    2f65c56 regexec.c: Handle sharp s in middle of backref
    6e326e8 regcomp.c: Handle more cases of tricky fold chars
    b1e3e56 regexec.c: Give context for ANYOFV call
    4e8910e regexec.c: Give context for ANYOFV call
    a5d74e1 regcomp.c: Remove special handling for U+00DF
    8e3094e regcomp.c: tell regexec more about multi-char folds
    d18bf9d regcomp.c: Synthetic start class should include ord >255 folds
    43322ea regcomp.c: Be more precise about ANYOF matching flag
    4c9daa0 regcomp.c: Put two static functions in embed.fnc
    17a3df4 Fix up \cX for 5.14
    0cd14f2 Minor bugfixes to Socket::getaddrinfo
    d1de25c Add supplied tests from perl #83194.
    33d9abf code points above 0x7fff_ffff problematic
    d764b54 Add initial inversion list object
    2f833f5 regcomp.c: Generate different property for /i matching
    6c29054 bmodpow() fails when GMP library is used.

There I have not included commits by Father Chrysostomos since he told
me that he "plan[s] to write perldelta entries for everything [he]
committed that is worthy of note.".

Later I noticed a mail by Karl Williamson where he said that he
"always[s] put in a delta if [he] think[s] one is needed". That along
with marking up the remaining as ignored turned the list of pending
commits into:

    $ perl Porting/git-deltatool --mode summary --type blocking
    Scanning for blocking commits since v5.13.9...
    $

We'll probably still need perldelta for"Fix up \cX for 5.14" and
related changes, but that can be done for the final 5.14 release.