handy.h: isIDFIRST_utf8() changed to use XIDStart
Previously this used a home-grown definition of an identifier start,
stemming from a bug in some early Unicode versions.  This led to some
problems, fixed by #74022.

But the home-grown solution did not track Unicode, and allowed for
characters, like marks, to begin words when they shouldn't.  This change
brings this macro into compliance with Unicode going-forward.