pp_regcomp: don't unset RXf_TAINTED
Currently at the end of this function, the flag RXf_TAINTED is set or
unset based the value of PL_tainting.

Skip the unsetting part. At the moment this is a noop, but in some future
variant, it might clear the flag that has already been set in some other
way.

In other words, be conservative.