In deprecate.pm, only cluck() once if the leaf isn't found in the caller stack.
Previously, if the expected leafname was not found in the caller stack, the
code was cluck, and then continue the loop with the next directories to scan.
However, this serves no useful purpose, as the loop changes nothing that causes
the failure condition, hence all that this would result in is a second cluck()
if the next scanned directories also happened to match. Now the core returns
immediately.