Skip t/re/pat_{advanced,rt_report}.t under minitest, as they use %- and %+
miniperl can't load Tie::Hash::NamedCapture, and as minitest is run if you
have build problems sufficiently severe that you can't bootstrap, it seems
unlikely that the advanced and obscure functionality of the regexp engine is
your most important concern. Hence better to skip it, than to try to bend it
to "portability".