Fix magic.t's expectation for $^X under minitest when the OS makes $^X absolute
Under minitest, t/perl is a symlink to ../perl. Test scripts are invoked with
./perl. When $Config{d_procselfexe} is undefined, $^X is still ./perl.
When it is defined, the OS resolves it to the absolute path .../miniperl.
Fix t/op/magic.t to be aware of this. Previously it was working for all cases
other than [$Config{d_procselfexe} is defined] and [make minitest] and
[perl does not exist].