In S_incpush, unixify libdir earlier.
This allows, for example, -I[.lib] to have Unix format appendages
added, such as "/buildcustomize.pl", "/sitecustomize.pl", etc.

It was previously only being converted to Unix syntax to allow the addition
of subdirectories, but the number of things that want to glue pieces onto
lib/ have multiplied over the years.