Ensure that chdir.t can find File::Spec and Cwd under minitest
As a working chdir is a key part of the build process, it would be counter-
productive to skip testing it if the build process failed, because it might be
cause of the problems. So add the source locations for Cwd and File::Spec in
dist/Cwd and dist/Cwd/lib respectively.