Move t/re/re.t to ext/re/t/re_funcs_u.t, so that it is not part of minitest.
The test file is for functions in the re:: namespace implemented in
universal.c, but needs to load re, which isn't built for minitest. As none of
these functions are used as part of the core's build process, seems best to
move it with all the other tests related to the re extension.