perldelta: Rearrange Core Enhancements into subsections
Also delete the perl -h thing, which is not noteworthy.

And capitalise headers consistently

Some of the items now under Regular Expressions are not specific to
regexps, but that should not matter, as the point is to create a
coherent document that people can read without getting lost.

And ‘Return value of C<delete $+{...}>’ has been ousted.